Rename remaining NodeProperties to NodeAppearance
