+++ examples/src/world-to-screen.ts
@@ -23,16 +23,16 @@ async function main() {
 
   let pickingNeedsUpdate = false;
   let pickedNode: number | undefined;
-  const shading = reveal_threejs.createDefaultShading({
+  const nodeAppearance: reveal_threejs.NodeAppearance = {
     color(treeIndex: number) {
       if (treeIndex === pickedNode) {
         return [0, 255, 255, 255];
       }
       return undefined;
     }
-  });
+  };
 
-  const cadNode = new CadNode(cadModel, { shading });
+  const cadNode = new CadNode(cadModel, { nodeAppearance });
   scene.add(cadNode);
 
   const renderer = new THREE.WebGLRenderer();
@@ -95,7 +95,7 @@ async function main() {
       } else {
         updateHtmlElements(0, 0, text);
       }
-      shading.updateNodes(updatedNodes);
+      cadNode.requestNodeUpdate(updatedNodes);
       pickingNeedsUpdate = true;
     }
   };
